+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `clear_extra` method
+- `set_extra_item` method
+- `get_extra_item` method
+- `remove_extra_item` method
 ### Changed
 - `_get_protocol` function modified
 - `decode_subscription` function blank line bug fixed